% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{bbox_to_polygon}
\alias{bbox_to_polygon}
\title{Convert a bounding box object into a polygon object}
\usage{
bbox_to_polygon(bbox)
}
\arguments{
\item{bbox}{bounding box object}
}
\value{
a polygon with the same oulines of the bounding box
}
\description{
Convert a bounding box object into a polygon object
}
