% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{coord_bbox}
\alias{coord_bbox}
\title{convenience function to cut off map at bounding box}
\usage{
coord_bbox(bbox)
}
\arguments{
\item{bbox}{bounding box}
}
\value{
a coord_sf object cutting the map view to the bounding box
}
\description{
convenience function to cut off map at bounding box
}
