% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{cov_vector_tiles}
\alias{cov_vector_tiles}
\title{Vector tiles for City of Vancouver + UBC
cached vector tiles for the City of Vancouver}
\usage{
cov_vector_tiles()
}
\value{
rmapzen type vector tile object
}
\description{
Vector tiles for City of Vancouver + UBC
cached vector tiles for the City of Vancouver
}
