% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{facet_transform_geos_by_group}
\alias{facet_transform_geos_by_group}
\title{Manipulates gemetries by group to make them suitable for faceting in a plot}
\usage{
facet_transform_geos_by_group(
  data,
  ...,
  scale_to_fill = FALSE,
  aspect_ratio = 1
)
}
\arguments{
\item{data}{datafrae with the geometries}

\item{...}{arguments to group by, these should be the same arguments used for faceting}

\item{scale_to_fill}{logical, will scale the geometries to fill each facet}

\item{aspect_ratio}{the aspect ratio for each facet, only used if scale_to_fill is `TRUE`}
}
\value{
a data frame with transformed geometries and `NA` crs
}
\description{
The data in each group will be transformed using a Lambert Equal Area Conic centred at the
group's geometry and centered. The projection preserves area, so this is useful if area comparisons
desired. Alternatively the `scale_to_fill``option can be used to scale the geometry to fill each
facet to the specified `aspect_ratio`.
}
