% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{file_to_s3}
\alias{file_to_s3}
\title{transfer file to aws s3}
\usage{
file_to_s3(path, s3_bucket, s3_path, content_type = NULL)
}
\arguments{
\item{path}{path to local file}

\item{s3_bucket}{s3 bucket name}

\item{s3_path}{s3 path in bucket, if it is a path component ending with a slash (`/`)
the basename of the input path will be appended}

\item{content_type}{mime type of the data, default is inferred from file extension}
}
\value{
upload result (boolean)
}
\description{
transfer file to aws s3
}
