% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_vetor_tiles.R
\name{geom_mapbox_roads}
\alias{geom_mapbox_roads}
\title{Adds a roads layer for the map}
\usage{
geom_mapbox_roads(
  ...,
  color = "black",
  width = 0.1,
  zoom = NULL,
  max_tiles = 5,
  transform = function(d) d[d$class \%in\% c("street", "primary", "secondary",
    "tertiary", "motorway"), ]
)
}
\arguments{
\item{...}{extra arguments}

\item{color}{color for the roads}

\item{width}{width of roads}

\item{zoom}{optional, zoom level for tiles}

\item{max_tiles}{Maximum number of tiles to query, overrides desired zoom level.}

\item{transform}{transform function to apply to the vector tile data, by default filter out ferry lines}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a roads layer for the map
}
