% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_vetor_tiles.R
\name{geom_mapbox_vector_tiles}
\alias{geom_mapbox_vector_tiles}
\title{Adds a vector tile layer for the map}
\usage{
geom_mapbox_vector_tiles(
  ...,
  type = c("water", "road", "building", "poi_label", "place_label", "structure",
    "transit_stop_label", "landuse"),
  zoom = NULL,
  max_tiles = 5,
  transform = function(d) d
)
}
\arguments{
\item{...}{extra arguments}

\item{type}{type of vector layer}

\item{zoom}{optional, zoom level for tiles}

\item{max_tiles}{Maximum number of tiles to query, overrides desired zoom level.}

\item{transform}{transform function to apply to the vector tile data}
}
\value{
a geom_sf object
}
\description{
Adds a vector tile layer for the map
}
