% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_vetor_tiles.R
\name{geom_mapbox_water}
\alias{geom_mapbox_water}
\title{Adds a water layer for the map}
\usage{
geom_mapbox_water(
  ...,
  fill = "lightblue",
  size = 0,
  zoom = NULL,
  max_tiles = 5,
  transform = function(d) d
)
}
\arguments{
\item{...}{extra arguments}

\item{fill}{fill for the water featuers, default is `lightblue`}

\item{size}{size of outline}

\item{zoom}{optional, zoom level for tiles}

\item{max_tiles}{Maximum number of tiles to query, overrides desired zoom level.}

\item{transform}{transform function to apply to the vector tile data}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a water layer for the map
}
