% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{geom_roads}
\alias{geom_roads}
\title{Adds a roads layer for the map}
\usage{
geom_roads(
  ...,
  color = "black",
  size = 0.1,
  nextzen_api_key = getOption("nextzen_API_key"),
  tile_size_px = NULL,
  transform = function(d) d[d$kind != "ferry", ]
)
}
\arguments{
\item{...}{extra arguments}

\item{color}{color for the roads}

\item{size}{size of roads}

\item{nextzen_api_key}{nextzen API key for vector tile data}

\item{tile_size_px}{tile size in pixels, may increase resolution of tile data}

\item{transform}{transform function to apply to the vector tile data, by default filter out ferry lines}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a roads layer for the map
}
