% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{geom_transit}
\alias{geom_transit}
\title{Adds a transit layer for the map}
\usage{
geom_transit(
  ...,
  color = "darkmagenta",
  size = 0.1,
  nextzen_api_key = getOption("nextzen_API_key"),
  tile_size_px = NULL,
  transform = function(d) d[d$kind \%in\% c("subway", "train"), ]
)
}
\arguments{
\item{...}{extra arguments}

\item{color}{color for the transit lines}

\item{size}{size of transit lines}

\item{nextzen_api_key}{nextzen API key for vector tile data}

\item{tile_size_px}{tile size in pixels, may increase resolution of tile data}

\item{transform}{transform function to apply to the vector tile data, by default filter out ferry lines}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a transit layer for the map
}
