% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{geom_vector_tiles}
\alias{geom_vector_tiles}
\title{Adds a vector tile layer for the map}
\usage{
geom_vector_tiles(
  ...,
  type = c("water", "buildings", "places", "transit", "pois", "boundaries", "roads",
    "earth", "landuse"),
  nextzen_api_key = getOption("nextzen_API_key"),
  tile_size_px = NULL,
  transform = function(d) d
)
}
\arguments{
\item{...}{extra arguments}

\item{type}{type of vector layer}

\item{nextzen_api_key}{nextzen API key for vector tile data}

\item{tile_size_px}{tile size in pixels, may increase resolution of tile data}

\item{transform}{transform function to apply to the vector tile data}
}
\value{
a geom_sf object
}
\description{
Adds a vector tile layer for the map
}
