% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{geom_water}
\alias{geom_water}
\title{Adds a water layer for the map}
\usage{
geom_water(
  ...,
  fill = "lightblue",
  width = 0,
  nextzen_api_key = getOption("nextzen_API_key"),
  tile_size_px = NULL,
  transform = function(d) d
)
}
\arguments{
\item{...}{extra arguments}

\item{fill}{fill for the water featuers, default is `lightblue`}

\item{width}{width of outline}

\item{nextzen_api_key}{nextzen API key for vector tile data}

\item{tile_size_px}{tile size in pixels, may increase resolution of tile data}

\item{transform}{transform function to apply to the vector tile data}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a water layer for the map
}
