% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_2016_census_hydro_layer}
\alias{get_2016_census_hydro_layer}
\title{2016 census hydro layer}
\usage{
get_2016_census_hydro_layer(
  cache_path = getOption("custom_data_path"),
  refresh = FALSE
)
}
\arguments{
\item{cache_path}{directory for caching the data, required, defaults to the `custom_data_path` option}

\item{refresh}{if true, download the data again and replace the cached version}
}
\value{
an sf object with the 2016 hydro layer
}
\description{
2016 census hydro layer
}
