% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_vetor_tiles.R
\name{get_mapbox_vector_tiles}
\alias{get_mapbox_vector_tiles}
\title{Get vector tile data}
\usage{
get_mapbox_vector_tiles(
  bbox,
  tileset_id = "mapbox.mapbox-streets-v8",
  zoom = NULL,
  max_tiles = 9,
  refresh = FALSE
)
}
\arguments{
\item{bbox}{bounding box for which to get vector tile data}

\item{tileset_id}{Mapbox tileset ID, default is "mapbox.mapbox-streets-v8"}

\item{zoom}{optional, zoom level for tiles}

\item{max_tiles}{Maximum number of tiles to query, overrides desired zoom level.}

\item{refresh}{results are cached for current session, optionally refresh cache}
}
\value{
a list of layers with vector tile data
}
\description{
expects Mapbox access token being available as MAPBOX_PUBLIC_TOKEN environment variable
}
