% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_metro_van_cut_geo_data_db}
\alias{get_metro_van_cut_geo_data_db}
\title{Metro Vancouver DA level geography cut down by MEtro Vancouevr land use data}
\usage{
get_metro_van_cut_geo_data_db(refresh = FALSE)
}
\arguments{
\item{refresh}{if true, refresh the data}
}
\description{
Metro Vancouver DA level geography cut down by MEtro Vancouevr land use data
}
