% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_metro_vancouver_land_use_data}
\alias{get_metro_vancouver_land_use_data}
\title{Metro Vancouver land use data}
\usage{
get_metro_vancouver_land_use_data(
  vintage = "2016.v2026",
  cache_path = getOption("custom_data_path"),
  refresh = FALSE
)
}
\arguments{
\item{vintage}{vintage of the data, valid parameters are `"2011"`, `"2016"`, `"2016.v20230517"`,
or any other string starting with `"2016."` like the default `"2016.v2026"` to label a snapshot of the current 2016 land use data}

\item{cache_path}{directory for caching the data}

\item{refresh}{if true, refresh the data}
}
\value{
an sf object with the metro vancouver land use data
}
\description{
Metro Vancouver land use data
}
