% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_statcan_geographic_attribute_data}
\alias{get_statcan_geographic_attribute_data}
\title{Geographic attribute data with DA representative points (2016 census version)}
\usage{
get_statcan_geographic_attribute_data(
  cache_path = getOption("custom_data_path"),
  refresh = FALSE
)
}
\arguments{
\item{cache_path}{directory for caching the data, required, defaults to the `custom_data_path` option}

\item{refresh}{if true, download the data again and replace the cached version}
}
\value{
an sf object with the representative point data
}
\description{
Geographic attribute data with DA representative points (2016 census version)
}
