% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_yvr_frequent_transit_network}
\alias{get_yvr_frequent_transit_network}
\title{Frequent transit layer}
\usage{
get_yvr_frequent_transit_network(refresh = FALSE)
}
\description{
Frequent transit layer
}
