% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{lambert_conformal_conic_at}
\alias{lambert_conformal_conic_at}
\title{get crs object for Lambert Conformal Conic projection centered at the input data}
\usage{
lambert_conformal_conic_at(data, center = NULL)
}
\arguments{
\item{data}{sf object}

\item{center}{optional centre point for the crs}
}
\value{
crs object
}
\description{
get crs object for Lambert Conformal Conic projection centered at the input data
}
