% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{metro_van_bbox}
\alias{metro_van_bbox}
\title{Vector tiles for Metro Vancouver}
\usage{
metro_van_bbox(clipped = TRUE)
}
\arguments{
\item{clipped}{should part of Nroth Shore mountains be clipped?}
}
\description{
Vector tiles for Metro Vancouver
}
