% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{parquet_to_r2}
\alias{parquet_to_r2}
\title{transfer parquet files to Cloudflare R2}
\usage{
parquet_to_r2(path, r2_bucket, r2_path = "")
}
\arguments{
\item{path}{path to local parquet file or directory with (partitioned) parquet files}

\item{r2_bucket}{R2 bucket name}

\item{r2_path}{path in bucket. If `path` is a file and `r2_path` is a path component ending with a slash (`/`)
the basename of the input path will be appended. If `path` is a directory `r2_path` is the path
under which the content of the directory gets placed.}
}
\value{
(invisibly) the paths in the bucket of the uploaded files
}
\description{
Uploads a single parquet file, or all parquet files in a directory of (partitioned) parquet files,
keeping the directory structure. The bucket has to exist, use `create_r2_bucket` to create it.
Expects the R2 account id and the access key id and secret access key
of an R2 API token to be available as `R2_ACCOUNT_ID`, `R2_ACCESS_KEY_ID` and `R2_SECRET_ACCESS_KEY`
environment variables.
}
