% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{pretty_cut}
\alias{pretty_cut}
\title{Cut values with pretty labels}
\usage{
pretty_cut(
  values,
  breaks,
  format = function(d) d,
  binding = "to",
  spacing = " ",
  under_text = "<",
  over_text = ">",
  ...
)
}
\arguments{
\item{values}{a numeric vector which is to be converted to a factor by cutting.}

\item{breaks}{a numeric vector of two or more unique cut points}

\item{format}{a function specifying how to format values in the labels}

\item{binding}{a string binding two values in the label}

\item{spacing}{a spacing string between elements in the label}

\item{under_text}{a string taking the space of the lower value and binding if the lower value is `-Inf`}

\item{over_text}{a string taking the space of the upper value and binding if the upper value is `Inf`}

\item{...}{additional arguments passed to `cut`}
}
\value{
a factor with the cut values and pretty labels
}
\description{
Cut values with pretty labels
}
