% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{roads_layer}
\alias{roads_layer}
\title{Adds a roads lay for the map}
\usage{
roads_layer(
  bbox,
  color = "black",
  size = 0.1,
  tile_size_px = NULL,
  transform = function(d) d
)
}
\arguments{
\item{bbox}{bounding box for the layer}

\item{color}{color for the outline}

\item{size}{size of outline}

\item{tile_size_px}{tile size in pixels, may increase resolution of tile data}

\item{transform}{transform function to apply to the vector tile data}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a roads lay for the map
}
