% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{sf_to_s3_gzip}
\alias{sf_to_s3_gzip}
\title{transfer sf object to gzipped geojson file on aws s3}
\usage{
sf_to_s3_gzip(data, s3_bucket, s3_path, precision = 7)
}
\arguments{
\item{data}{sf object}

\item{s3_bucket}{s3 bucket name}

\item{s3_path}{s3 path in bucket}

\item{precision}{number of digits to write geojson}
}
\value{
upload result (boolean)
}
\description{
transfer sf object to gzipped geojson file on aws s3
}
