% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mm.R
\name{theme_mm}
\alias{theme_mm}
\title{MountainMath theme}
\usage{
theme_mm(..., background_colour = "#F8F4F0")
}
\arguments{
\item{...}{Additional parameters passed to theme()}

\item{background_colour}{Background colour for plot and panel}
}
\value{
A theme to be added to a ggplot object
}
\description{
MountainMath theme
}
