% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{water_layer}
\alias{water_layer}
\title{Adds a water lay for the map}
\usage{
water_layer(
  bbox,
  fill = "lightblue",
  color = NA,
  size = 1,
  tile_size_px = NULL,
  transform = function(d) d
)
}
\arguments{
\item{bbox}{bounding box for the layer}

\item{fill}{fill color}

\item{color}{color for the outline}

\item{size}{size of outline}

\item{tile_size_px}{tile size in pixels, may increase resolution of tile data}

\item{transform}{transform function to apply to the vector tile data}
}
\value{
a geom_sf object with the layer
}
\description{
Adds a water lay for the map
}
