% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{cov_bbox}
\alias{cov_bbox}
\title{Vector tiles for City of Vancouver}
\usage{
cov_bbox(include_ubc = TRUE)
}
\arguments{
\item{include_ubc}{include UBC area in boundng box}
}
\value{
a bounding box
}
\description{
Vector tiles for City of Vancouver
}
