% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocode}
\alias{geocode}
\title{Use BC Geocoder to geocode addresses. This is slow, goes through address list one by one}
\usage{
geocode(data, address_field = "addressString")
}
\arguments{
\item{data}{data frame with rows to be geocoded}

\item{address_field}{column in data frame that contains the address string}
}
\value{
data frame with `X`, `Y` columns for longitude and latitude, as well as
matching `score`, `matchPrecision` and cleaned `usedAddressString`
}
\description{
Use BC Geocoder to geocode addresses. This is slow, goes through address list one by one
}
