% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_2016_census_fsa_geos}
\alias{get_2016_census_fsa_geos}
\title{2016 census FSA geography}
\usage{
get_2016_census_fsa_geos(
  cache_path = getOption("custom_data_path"),
  refresh = FALSE
)
}
\arguments{
\item{cache_path}{directory for caching the data}

\item{refresh}{if true, refresh the data}
}
\value{
an sf object with the 2016 FSA geography
}
\description{
2016 census FSA geography
}
