% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_2016_census_hydro_layer}
\alias{get_2016_census_hydro_layer}
\title{2016 census hydro layer}
\usage{
get_2016_census_hydro_layer(refresh = FALSE)
}
\arguments{
\item{refresh}{if true, refresh the data}
}
\description{
2016 census hydro layer
}
