% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_helpers.R
\name{get_metro_vancouver_land_use_data}
\alias{get_metro_vancouver_land_use_data}
\title{Metro Vancouver land use data (2011 version)}
\usage{
get_metro_vancouver_land_use_data(refresh = FALSE)
}
\description{
Metro Vancouver land use data (2011 version)
}
