% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tile_helpers.R
\name{get_vector_tiles}
\alias{get_vector_tiles}
\title{Get vector tile data, expects option variable nextzen_API_key to be set}
\usage{
get_vector_tiles(bbox, width = NULL, height = NULL, refresh = FALSE)
}
\arguments{
\item{bbox}{bounding box for which to get vector tile data}

\item{width}{optional, width in pixels for the display image}

\item{height}{optional, width in height for the display image}

\item{refresh}{results are cached for current session, optionally refresh cache}

\item{bbox}{bounding box for which to get vector tile data}
}
\value{
a list of layers with vector tile data
}
\description{
Get vector tile data, expects option variable nextzen_API_key to be set
}
